refactor: move ReferralHeroSection
