+++ .gitignore
@@ -139,3 +139,4 @@ dist
 *.njsproj
 *.sln
 *.sw?
+.cursor/*
