--- src/components/ReferralHeroSection/index.ts
@@ -1,1 +1,0 @@
-export { default as ReferralHeroSection } from './ReferralHeroSection.container'
