+++ src/components/Referrals/ReferralHeroSection/ReferralHeroSection.types.ts
@@ -1,0 +1,8 @@
+import { View } from '../../../utils/view'
+
+export type Props = {
+  isLoading?: boolean
+  error?: string | null
+  view?: View
+  profileAddress: string
+}
