+++ src/components/Referrals/ReferralHeroSection/index.ts
@@ -1,0 +1,1 @@
+export { ReferralHeroSection } from './ReferralHeroSection'
